В настройки компании добавлена валюта по умолчанию, произведен рефакторинг карточки товара, убрана опция продукция

+++ GriB.Client.App/Content/js/app/controller/setting/editor/product.ts
@@ -1,5 +1,6 @@
 ﻿import vars = require('app/common/variables');
 import utils = require('app/common/utils');
+import ctrl = require('app/common/basecontrol');
 import edit = require('app/controller/setting/editor/editor');
 
 export namespace Controller.Setting.Editor {
@@ -106,9 +107,9 @@ export namespace Controller.Setting.Editor {
         private controlType: JQuery;
         private imgDialog: any;
         private controlPhoto: JQuery;
-        private categoryList: JQuery;
-        private currencyList: JQuery;
-        private unitList: JQuery;
+        private categoryControl: ctrl.Control.ReferenceFieldControl;
+        private unitControl: ctrl.Control.ReferenceFieldControl;
+        private currencyControl: ctrl.Control.ReferenceFieldControl;
         private compositionRows: JQuery;
         private rightRows: JQuery;
         private btnAddComposition: JQuery;
@@ -116,13 +117,16 @@ export namespace Controller.Setting.Editor {
         public ViewInit(view: JQuery): boolean {
             this.imgDialog = view.find("#editor-view-image-input");
             this.controlPhoto = view.find("#editor-view-product-photo");
-            this.categoryList = view.find("#editor-view-category-list");
-            this.unitList = view.find("#editor-view-product-unit");
-            this.currencyList = view.find("#editor-view-product-currency");
             this.compositionRows = view.find("#product-composition-rows");
             this.rightRows = view.find("#product-rigths-rows");
             this.controlType = view.find('#editor-view-product-type');
 
+            this.categoryControl = new ctrl.Control.ReferenceFieldControl();
+            this.categoryControl.InitControl(view.find("#editor-view-product-category-row"), "editor-view-product-category", "editModel.category", "editModel.category.name", vars._statres("label$includedincategory"), 'setting/card/category', this.Model);
+            this.unitControl = new ctrl.Control.ReferenceFieldControl();
+            this.unitControl.InitControl(view.find("#editor-view-product-unit-row"), "editor-view-product-unit", "editModel.unit", "editModel.unit.name", vars._statres("label$unit"), 'setting/card/unit', this.Model);
+            this.currencyControl = new ctrl.Control.ReferenceFieldControl();
+            this.currencyControl.InitControl(view.find("#editor-view-product-currency-row"), "editor-view-product-currency", "editModel.currency", "editModel.currency.name", vars._statres("label$currency"), 'setting/card/currency', this.Model);
 
             view.find("#editor-view-product-name").characterCounter();
             view.find("#editor-view-product-description").characterCounter();
@@ -145,6 +149,13 @@ export namespace Controller.Setting.Editor {
             let onUpolad = $.proxy(this.uploudImageClick, this);
             this.imgDialog.bind("change", onUpolad);
 
+            if (this.categoryControl) {
+                this.categoryControl.createEvents();
+                this.categoryControl.SelectValue = $.proxy(this.selectCategoryValue, this);
+            }
+            if (this.unitControl) this.unitControl.createEvents();
+            if (this.currencyControl) this.currencyControl.createEvents();
+
             this.AddPhotoButtonClick = this.createTouchClickEvent("editor-view-product-addphoto", this.addPhotoButtonClick);
             this.Model.bind("change", $.proxy(this.changeModel, this));
         }
@@ -154,16 +165,24 @@ export namespace Controller.Setting.Editor {
             this.rightRows.unbind();
             this.Model.unbind("change");
             this.destroyTouchClickEvent("editor-view-product-addphoto", this.AddPhotoButtonClick);
-            if (this.btnAddComposition)
-                this.destroyTouchClickEvent(this.btnAddComposition, this.addCompositionButtonClick);
-            if (this.btnRemoveComposition)
-                this.destroyTouchClickEvent(this.btnRemoveComposition, this.removeCompositionButtonClick);
+
+            if (this.currencyControl) this.currencyControl.destroyEvents();
+            if (this.unitControl) this.unitControl.destroyEvents();
+            if (this.categoryControl) this.categoryControl.destroyEvents();
+
+            if (this.btnAddComposition) this.destroyTouchClickEvent(this.btnAddComposition, this.addCompositionButtonClick);
+            if (this.btnRemoveComposition) this.destroyTouchClickEvent(this.btnRemoveComposition, this.removeCompositionButtonClick);
             this.imgDialog.unbind();
             super.destroyEvents();
 
         }
 
-      
+        private selectCategoryValue(value: Interfaces.Model.IBaseModel) {
+            let valueCategory: any = value;
+            if (valueCategory) {
+                this.Model.set("editModel.category", { id: valueCategory.id, name: (utils.isNullOrEmpty(valueCategory.parentname) === true ? "" : valueCategory.parentname + '->') + valueCategory.name });
+            }
+        }
         private changeModel(e: any): void {
             if (e.field === "editModel.type") {
                 let model: Interfaces.Model.IProduct = this.EditorModel;
@@ -208,57 +227,12 @@ export namespace Controller.Setting.Editor {
         protected afterLoad(responseData?: any): void {
             super.afterLoad(responseData);
             this.controlPhoto.css("backgroundImage", "url(" + this.EditorModel.photo + ")");
-            this.setupListCategory(responseData);
-            this.setupListUnit(responseData);
-            this.setupListCurrencies(responseData);
             this.changeModel({ field: "editModel.type" });
             this.setupTableComposition();
 
             let model: Interfaces.Model.IProduct = this.EditorModel;
             let data: Interfaces.Model.ISalePointAccessModel[] = model.accesssalepoints;
             this.setupTableAccess(this.rightRows, data);
-        }
-
-        private setupListCategory(responseData: any) {
-            let html: string = '';
-            let categories = responseData.categories;
-            html += ' <option value="0"' + (0 === this.EditorModel.pid ? ' selected' : '') + '>' + vars._statres("label$categorynotspecified") + '</option>';
-            if (categories && categories.length > 0) {
-                for (let i = 0, icount = categories.length; i < icount; i++)
-                    html += ' <option value="' + categories[i].id + '"' + (categories[i].id === this.EditorModel.pid ? ' selected' : '') + '>' + categories[i].name + '</option>';
-                this.categoryList.html(html);
-            }
-            else
-                this.categoryList.html('');
-            this.categoryList.formSelect();
-        }
-
-        private setupListUnit(responseData: any) {
-            let html: string = '';
-            let units = responseData.units;
-            html += ' <option value="0"' + (0 === this.EditorModel.pid ? ' selected' : '') + '>' + vars._statres("label$unitnotspecified") + '</option>';
-            if (units && units.length > 0) {
-                for (let i = 0, icount = units.length; i < icount; i++)
-                    html += ' <option value="' + units[i].id + '"' + (units[i].id === this.EditorModel.unit ? ' selected' : '') + '>' + units[i].name + '</option>';
-                this.unitList.html(html);
-            }
-            else
-                this.unitList.html('');
-            this.unitList.formSelect();
-        }
-
-        private setupListCurrencies(responseData: any) {
-            let html: string = '';
-            let currencies = responseData.currencies;
-            html += ' <option value="0"' + (0 === this.EditorModel.pid ? ' selected' : '') + '>' + vars._statres("label$currencynotspecified") + '</option>';
-            if (currencies && currencies.length > 0) {
-                for (let i = 0, icount = currencies.length; i < icount; i++)
-                    html += ' <option value="' + currencies[i].id + '"' + (currencies[i].id === this.EditorModel.currency ? ' selected' : '') + '>' + currencies[i].code + (utils.isNullOrEmpty(currencies[i].name) ? '' : ' - ') + currencies[i].name + '</option>';
-                this.currencyList.html(html);
-            }
-            else
-                this.currencyList.html('');
-            this.currencyList.formSelect();
         }
 
         private setupTableComposition(): void {
@@ -390,7 +364,7 @@ export namespace Controller.Setting.Editor {
 
         protected getSaveModel(): Interfaces.Model.IEditorModel {
             let model: Interfaces.Model.IProduct = this.EditorModel;
-            let catg = this.categoryList.val();
+            let catg = this.Model.get("editModel.category.id");
             model.pid = +catg; //(catg && catg.length > 0 ? +catg[0] : 0);
 
             return model;
